+++ plantcv/geospatial/analyze/__init__.py
@@ -1,6 +1,7 @@
 from plantcv.geospatial.analyze.coverage import coverage
 from plantcv.geospatial.analyze.dsm import height_percentile
 from plantcv.geospatial.analyze.spectral import spectral_index
+from plantcv.geospatial.analyze.dsm import height_subtraction
 
 
-__all__ = ["coverage", "height_percentile", "spectral_index"]
+__all__ = ["coverage", "height_percentile", "height_subtraction", "spectral_index"]